Changed file structure, particularly includes
This commit changes how the includes are being done in the current
library.

This now works like this:
- There is now a public API for the library in include/, this header
file located under "Ark/String.hpp" is the one that is going to be
installed on machines when installing this library has a standalone and
the one we will be using when compiling from a different project

The old way used this include in the source file which is not a good way
to do this. Instead, the inl and the .hpp file with private
attributes are in the src directory. Everything that is not supposed to
be used by the user must not be exposed and installed but must be
defined in some classes or header in the src/ directory. This is the
case now.

Now build from the src/ directory the library and then in future commits
will link it against test within the test executable in test/